requirements.txt
pydub
python-telegram-bot
pyyaml
spleeter
